# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Internal Developer Portal API
  description: >
    Service catalog API: teams, services, dependencies and API specs.
    This spec is the single source of truth — the Go server interfaces and the
    TypeScript client are both generated from it.
  version: 0.1.0
  contact:
    name: Otso Heinämäki
  license:
    name: MIT
    url: https://opensource.org/license/mit
servers:
  - url: /api/v1
    description: Mounted under /api/v1 by the server

# Entra ID JWT bearer auth; enforced from week 2 (dev-mode fallback locally).
security:
  - bearerAuth: []

tags:
  - name: services
    description: Service catalog entries
  - name: teams
    description: Owning teams
  - name: auth
    description: The signed-in user and their roles

paths:
  /services:
    get:
      operationId: ListServices
      tags: [services]
      summary: List services
      # TODO(week 3): these filters are validated against this spec but the
      # handler still ignores them; they land with Postgres FTS.
      parameters:
        - name: team
          in: query
          description: Filter by owning team slug
          schema:
            type: string
        - name: lifecycle
          in: query
          description: Filter by lifecycle stage
          schema:
            $ref: "#/components/schemas/Lifecycle"
        - name: tag
          in: query
          description: Filter by tag name
          schema:
            type: string
        - name: q
          in: query
          description: Free-text search over name and description
          schema:
            type: string
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: CreateService
      tags: [services]
      summary: Register a new service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceCreate"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/ErrorResponse"

  /services/{serviceId}:
    parameters:
      - name: serviceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: GetService
      tags: [services]
      summary: Get a service by id
      responses:
        "200":
          description: The service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/ErrorResponse"
    put:
      operationId: UpdateService
      tags: [services]
      summary: Update a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      operationId: DeleteService
      tags: [services]
      summary: Delete a service
      description: Requires the ADMIN role on the owning team.
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/ErrorResponse"

  /teams:
    get:
      operationId: ListTeams
      tags: [teams]
      summary: List teams
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/ErrorResponse"

  /me:
    get:
      operationId: GetCurrentUser
      tags: [auth]
      summary: The signed-in user and their per-team roles
      description: >
        Roles are read from the database on every call rather than carried in
        the token, so membership changes take effect without re-authenticating.
      responses:
        "200":
          description: The signed-in user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Microsoft Entra ID access token

  # One named response per status code, deliberately. oapi-codegen keeps a
  # per-operation set of already-used response refs: only the FIRST response in
  # an operation that points at a given component gets the embedded
  # `struct{ XJSONResponse }` form, and later ones degrade to a bare
  # `type Y Error`. Sharing a single ErrorResponse across statuses therefore
  # makes the generated Go shapes depend on which status codes an operation
  # happens to declare — adding a 401 silently reshapes its 404. Distinct refs
  # never collide, so every fixed-status error response stays embedded.
  responses:
    BadRequest:
      description: The request body or parameters are invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

    Forbidden:
      description: Authenticated, but lacking the required role on the owning team
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

    NotFound:
      description: No such resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

    Conflict:
      description: Conflicts with an existing resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

    # Used only by `default`, which always generates the {Body, StatusCode} form.
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    Lifecycle:
      type: string
      description: Lifecycle stage of a service
      enum: [production, beta, deprecated]

    Role:
      type: string
      description: Role held within a single team
      enum: [ADMIN, EDITOR, VIEWER]
      # Without this the generator emits bare constants, exporting ADMIN,
      # EDITOR and VIEWER from package api.
      x-enum-varnames: [RoleAdmin, RoleEditor, RoleViewer]

    TeamRole:
      type: object
      description: >
        A team the user belongs to, and their role in it. Team name and slug are
        denormalized so the UI can render a team picker from this alone.
      required: [teamId, teamSlug, teamName, role]
      properties:
        teamId:
          type: string
          format: uuid
        teamSlug:
          type: string
        teamName:
          type: string
        role:
          $ref: "#/components/schemas/Role"

    CurrentUser:
      type: object
      description: >
        The signed-in user. An empty teamRoles list still grants read access to
        the whole catalog — every authenticated user is an implicit viewer.
      required: [id, email, name, teamRoles]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        teamRoles:
          type: array
          items:
            $ref: "#/components/schemas/TeamRole"

    TeamRef:
      type: object
      description: Reference to the owning team
      required: [id, name, slug]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string

    Team:
      type: object
      required: [id, name, slug, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        createdAt:
          type: string
          format: date-time

    TeamList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Team"

    Service:
      type: object
      required: [id, name, slug, lifecycle, team, tags, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        lifecycle:
          $ref: "#/components/schemas/Lifecycle"
        repoUrl:
          type: string
          format: uri
        runbookUrl:
          type: string
          format: uri
        team:
          $ref: "#/components/schemas/TeamRef"
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ServiceCreate:
      type: object
      required: [name, slug, teamId, lifecycle]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        slug:
          type: string
          pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"
          maxLength: 120
        description:
          type: string
          maxLength: 2000
        lifecycle:
          $ref: "#/components/schemas/Lifecycle"
        repoUrl:
          type: string
          format: uri
        runbookUrl:
          type: string
          format: uri
        teamId:
          type: string
          format: uuid
        tags:
          type: array
          items:
            type: string

    ServiceUpdate:
      type: object
      required: [name, lifecycle, teamId]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 2000
        lifecycle:
          $ref: "#/components/schemas/Lifecycle"
        repoUrl:
          type: string
          format: uri
        runbookUrl:
          type: string
          format: uri
        teamId:
          type: string
          format: uuid
        tags:
          type: array
          items:
            type: string

    ServiceList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Service"

    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable message